Remove DummyFileSystem hack for symfony 5.x, fix phpstan / rector issues
