+++ devTools/phpstan-src.neon
@@ -29,16 +29,6 @@ parameters:
             path: '../src/Logger/Html/StrykerHtmlReportBuilder.php'
             message: '#return type with generic class ArrayObject does not specify its types\: TKey, TValue#'
         -
-            message: "#^Instantiated class Infection\\\\FileSystem\\\\DummyFileSystem not found\\.$#"
-            count: 1
-            path: ../src/Container.php
-
-        -
-            message: "#^Parameter \\#5 \\$fileSystem of class Infection\\\\Process\\\\Runner\\\\MutationTestingRunner constructor expects Symfony\\\\Component\\\\Filesystem\\\\Filesystem, Infection\\\\FileSystem\\\\DummyFileSystem\\|Symfony\\\\Component\\\\Filesystem\\\\Filesystem given\\.$#"
-            count: 1
-            path: ../src/Container.php
-
-        -
             message: "#^ArrayObject\\<\\*NEVER\\*, \\*NEVER\\*\\> does not accept array\\<string, array\\|string\\>\\.$#"
             count: 1
             path: ../src/Logger/Html/StrykerHtmlReportBuilder.php
